Replace provision k3d with alpha impl
