--- cmd/kyma/alpha/provision/cmd.go
@@ -1,14 +1,0 @@
-package provision
-
-import (
-	"github.com/spf13/cobra"
-)
-
-//NewCmd creates a new provision command
-func NewCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "provision",
-		Short: "Provisions a cluster for Kyma installation.",
-	}
-	return cmd
-}
