+++ cmd/kyma/provision/k3d/cmd.go
@@ -22,7 +22,6 @@ type command struct {
 
 //NewCmd creates a new k3d command
 func NewCmd(o *Options) *cobra.Command {
-
 	c := command{
 		Command: cli.Command{Options: o.Options},
 		opts:    o,
@@ -30,7 +29,7 @@ func NewCmd(o *Options) *cobra.Command {
 
 	cmd := &cobra.Command{
 		Use:     "k3d",
-		Short:   "Provisions a Kubernetes cluster based on k3d v4.",
+		Short:   "Provisions a Kubernetes cluster based on k3d v5.",
 		Long:    `Use this command to provision a k3d-based Kubernetes cluster for Kyma installation.`,
 		RunE:    func(_ *cobra.Command, _ []string) error { return c.Run() },
 		Aliases: []string{"k"},
@@ -38,11 +37,12 @@ func NewCmd(o *Options) *cobra.Command {
 
 	cmd.Flags().StringVar(&o.Name, "name", "kyma", `Name of the Kyma cluster`)
 	cmd.Flags().IntVar(&o.Workers, "workers", 1, "Number of worker nodes (k3d agents)")
-	cmd.Flags().StringSliceVarP(&o.ServerArgs, "server-arg", "s", []string{}, "One or more arguments passed to the Kubernetes API server (e.g. --server-arg='--alsologtostderr')")
-	cmd.Flags().StringSliceVarP(&o.AgentArgs, "agent-arg", "a", []string{}, "One or more arguments passed to the k3d agent command on agent nodes (e.g. --agent-arg='--alsologtostderr')")
+	cmd.Flags().StringSliceVarP(&o.K3sArgs, "k3s-arg", "s", []string{}, "One or more arguments passed from k3d to the k3s command (format: ARG@NODEFILTER[;@NODEFILTER])")
 	cmd.Flags().DurationVar(&o.Timeout, "timeout", 5*time.Minute, `Maximum time for the provisioning. If you want no timeout, enter "0".`)
 	cmd.Flags().StringSliceVarP(&o.K3dArgs, "k3d-arg", "", []string{}, "One or more arguments passed to the k3d provisioning command (e.g. --k3d-arg='--no-rollback')")
-	cmd.Flags().StringVarP(&o.KubernetesVersion, "kube-version", "k", "1.20.11", "Kubernetes version of the cluster")
+	cmd.Flags().StringVarP(&o.MinorKubernetesVersion, "kube-version", "k", "1.20", "Minor Kubernetes version of the cluster")
+	cmd.Flags().StringSliceVar(&o.UseRegistry, "registry-use", []string{}, "Connect to one or more k3d-managed registries. Kyma automatically creates a registry for serverless images.")
+	cmd.Flags().StringVar(&o.RegistryPort, "registry-port", "5001", "Specify the port on which the k3d registry will be exposed")
 	cmd.Flags().StringSliceVarP(&o.PortMapping, "port", "p", []string{"80:80@loadbalancer", "443:443@loadbalancer"}, "Map ports 80 and 443 of K3D loadbalancer (e.g. -p 80:80@loadbalancer -p 443:443@loadbalancer)")
 	return cmd
 }
@@ -57,14 +57,17 @@ func (c *command) Run() error {
 	}
 
 	var err error
+	var registryURL string
 
 	k3dClient := k3d.NewClient(k3d.NewCmdRunner(), k3d.NewPathLooker(), c.opts.Name, c.opts.Verbose, c.opts.Timeout)
 
 	if err = c.verifyK3dStatus(k3dClient); err != nil {
 		return err
 	}
-
-	if err = c.createK3dCluster(k3dClient); err != nil {
+	if registryURL, err = c.createK3dRegistry(k3dClient); err != nil {
+		return err
+	}
+	if err = c.createK3dCluster(k3dClient, registryURL); err != nil {
 		return err
 	}
 	return nil
@@ -73,7 +76,7 @@ func (c *command) Run() error {
 //Verifies if k3d is properly installed and pre-conditions are fulfilled
 func (c *command) verifyK3dStatus(k3dClient k3d.Client) error {
 	s := c.NewStep("Verifying k3d status")
-	if err := k3dClient.VerifyStatus(false); err != nil {
+	if err := k3dClient.VerifyStatus(true); err != nil {
 		s.Failure()
 		return err
 	}
@@ -81,7 +84,14 @@ func (c *command) verifyK3dStatus(k3dClient k3d.Client) error {
 	s.LogInfo("Checking if port flags are valid")
 	ports, err := extractPortsFromFlag(c.opts.PortMapping)
 	if err != nil {
-		return errors.Wrap(err, fmt.Sprintf("Could not extract host ports from %s", c.opts.PortMapping))
+		return errors.Wrapf(err, "Could not extract host ports from %s", c.opts.PortMapping)
+	}
+
+	s.LogInfo("Checking if k3d registry of previous kyma installation exists")
+	registryExists, err := k3dClient.RegistryExists()
+	if err != nil {
+		s.Failure()
+		return err
 	}
 
 	s.LogInfo("Checking if k3d cluster of previous kyma installation exists")
@@ -97,14 +107,33 @@ func (c *command) verifyK3dStatus(k3dClient k3d.Client) error {
 			return fmt.Errorf("User decided not to remove the existing k3d cluster")
 		}
 
+		if registryExists {
+			if err := k3dClient.DeleteRegistry(); err != nil {
+				s.Failure()
+				return err
+			}
+			s.LogInfo("Deleted k3d registry of previous kyma installation")
+		}
+
 		if err := k3dClient.DeleteCluster(); err != nil {
 			s.Failure()
 			return err
 		}
 		s.LogInfo("Deleted k3d cluster of previous kyma installation")
-	} else if err := allocatePorts(ports...); err != nil {
-		s.Failure()
-		return errors.Wrap(err, "Port cannot be allocated")
+
+	} else {
+		if err := allocatePorts(ports...); err != nil {
+			s.Failure()
+			return errors.Wrap(err, "Port cannot be allocated")
+		}
+		if registryExists {
+			// only registry exists
+			if err := k3dClient.DeleteRegistry(); err != nil {
+				s.Failure()
+				return err
+			}
+			s.LogInfo("Deleted k3d registry of previous kyma installation")
+		}
 	}
 
 	s.Successf("k3d status verified")
@@ -119,22 +148,36 @@ func (c *command) PromptUserToDeleteExistingCluster() bool {
 	return c.opts.NonInteractive || answer
 }
 
+func (c *command) createK3dRegistry(k3dClient k3d.Client) (string, error) {
+	s := c.NewStep("Create k3d registry")
+
+	registryURL, err := k3dClient.CreateRegistry(c.opts.RegistryPort)
+	if err != nil {
+		s.Failuref("Could not create k3d registry")
+		return "", err
+	}
+	s.Successf("Created k3d registry '%s'", registryURL)
+	return registryURL, nil
+}
+
 //Create a k3d cluster
-func (c *command) createK3dCluster(k3dClient k3d.Client) error {
+func (c *command) createK3dCluster(k3dClient k3d.Client, registryURL string) error {
 	s := c.NewStep(fmt.Sprintf("Create K3d cluster '%s'", c.opts.Name))
+
+	c.opts.UseRegistry = append(c.opts.UseRegistry, registryURL)
 
 	settings := k3d.CreateClusterSettings{
 		Args:              parseK3dArgs(c.opts.K3dArgs),
-		KubernetesVersion: c.opts.KubernetesVersion,
+		KubernetesVersion: c.opts.MinorKubernetesVersion,
 		PortMapping:       c.opts.PortMapping,
 		Workers:           c.opts.Workers,
-		V4Settings: k3d.V4CreateClusterSettings{
-			ServerArgs: c.opts.ServerArgs,
-			AgentArgs:  c.opts.AgentArgs,
+		V5Settings: k3d.V5CreateClusterSettings{
+			K3sArgs:     c.opts.K3sArgs,
+			UseRegistry: c.opts.UseRegistry,
 		},
 	}
 
-	err := k3dClient.CreateCluster(settings, false)
+	err := k3dClient.CreateCluster(settings, true)
 	if err != nil {
 		s.Failuref("Could not create k3d cluster '%s'", c.opts.Name)
 		return err
